requirements.txt
absl-py==1.2.0
amqp==5.1.1
aniso8601==7.0.0
apturl==0.5.2
asgiref==3.5.2
astunparse==1.6.3
bcrypt==3.2.0
billiard==3.6.4.0
black==22.6.0
blinker==1.4
boto3==1.24.46
botocore==1.27.46
Brlapi==0.8.3
cachetools==5.2.0
celery==5.2.7
certifi==2020.6.20
chardet==4.0.0
ci-info==0.3.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
configparser==5.2.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dj-rest-auth==2.2.5
Django==4.0.6
django-allauth==0.51.0
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-filter==22.1
django-ipware==4.0.2
django-storages==1.13.1
django-structlog==3.0.1
django-timezone-field==5.0
djangogrpcframework==0.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
duplicity==0.8.21
environ==1.0
etelemetry==0.3.0
fasteners==0.14.1
filelock==3.6.0
fitz==0.0.1.dev2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.47.0
grpcio-tools==1.47.0
h5py==3.7.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isodate==0.6.1
jeepney==0.7.1
jmespath==1.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.1
lockfile==0.12.2
looseversion==1.0.1
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
messagebird==2.1.0
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==0.4.3
netifaces==0.11.0
networkx==2.8.5
nibabel==4.0.1
nipype==1.8.3
numpy==1.23.1
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
paramiko==2.9.3
pathlib==1.0.1
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.12.4
prov==2.0.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
pydot==1.4.2
PyGObject==3.42.1
PyJWT==2.4.0
pymacaroons==0.13.0
PyMuPDF==1.20.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytesseract==0.3.9
python-apt==2.3.0+ubuntu2.1
python-crontab==2.6.0
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.20.0
python-environ==0.4.54
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
pyxnat==1.4
PyYAML==5.4.1
rdflib==6.2.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
Rx==1.6.1
s3transfer==0.6.0
scipy==1.9.0
SecretStorage==3.3.1
simplejson==3.17.6
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
stripe==2.63.0
structlog==22.1.0
systemd-python==234
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
text-unidecode==1.3
tomli==2.0.1
traits==6.3.2
typing_extensions==4.3.0
tzdata==2022.1
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.2.0
whitenoise==6.2.0
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
